Proxies just won't die.
I have had to re-introduce the proxify feature because
JScript seems to play fast-and-loose with object property
casings. For example:

  var wsh = new ActiveXObject("wscript.shell");
  wsh.EnViROnMeNT("SYSTEM");
  //  ^^^^^^^^^^^
  //  Legal.

Proxies are now used to drop the case of all methods and
properties.  It's messy, and hacky, and I'm sure this will
come back to bite me soon, but this is the best solution I
have right now.