Added 'convert_to' functions
Both Binary and Text streams now have the ability to convert between
eachother.  This isn't intended for general use, but the ADODBStream
supports converting between the two, so these methods (and tests) were
added.

Also updated the VFS default encoding from 'ascii' to 'utf16le' to
fall in line with Win7 and its default charcode of "Unicode".  This is
not a full implementation of charcodes.