First commit of rewritten ADODB Stream
ADODB Stream is being rewritten to use the underlying Text and Binary
stream instances.  This is the first iteration of that.  Many of the
method comments which were lost as part of this commit will be brought
back -- but for speed and clarity, they're been removed while tests
are written.