Added new method: `put_buf'
This method will help both Text and Binary streams when writing new
data to their internal buffers.  It will handle expanding internal
buffers, as well as updating positions.  It is not charset aware, and
only understands buffers.  It's the job of calling-code to ensure
`.pos' is set appropriately when using BOMs.