Large updates, mostly surrounding .Mode
Significant update to the handling of the .Mode property, including
stream read/write permissions, and associated exception handling.

Also fixed a bug in the 'read_n_chars' method (TextStream) which would
crash if called without any value.  It will now fetch the rest...of the
stream if no arguments are provided.

Also did some housekeeping to replace all 'this_context' references
with 'ctx' for consistency, and updated the TODO list.