Added #Read Tests
Finished adding the #Read tests.  Frustratingly, #Read returns a value
which is on the otherside of a "COM bridge", meaning the type returned
is not accessible from within JScript.  Not much we can do about it
here.  For now, we just return a Buffer instance (something which does
not exist in JScript).  Needs more thought.