FSO Updates + Auto-Vivify and CFG
Updated the FileSystemObject as part of the ongoing implementation work
for this component.  Added 'autovivification' which is a means by which
folders are auto-created (if enabled) instead of throwing a 'path not
found' exception.  Highly inspired by Perl's approach to hash key
creation.

As part of the autoviv work, there is now a new method for fetching
both environment and configuration values, via the methods:

 - `get_env`
 - `get_cfg`

As maintenance on the system progresses, I'd like to phase-out direct
envenvironmental access and route all requests through either of these
functions.  Finally, the separation between CFG and ENV allows a
user-supsupplied config to drastically alter the behaviour of Construct.