Tidied ShortName code + Updated Wildcard.org
ShortName code and comments has been updated.  Still some work
remaining to correctly update the shortname table.

Also ripped out a lot of the wildcard.org file as much of it was wrong.
Have added links to all of the useful articles which will help when
writing the wildcard matcher.