Cleaned-up code and fixed bugs RE: symlinks
Added ntfs method `mkdirpSync'.  This exposed an issue where calling:

    mkdirpSync("/HelloWorld/foo");
    mkdirpSync("/HelloWorld/foo/bar");

On the second call we would create another link to "HelloWorld".  This
is now fixed, and mkdirpSync is safe to use.

Also renamed function parameters to indicate whether the path is an
external path ("C:\Foo\Bar") or an internal one ("/Foo/Bar").