Moved is_filename to wildcard module + tests
The is_filename function has been moved in to the wildcard module so that it
can be shared by both the VFS and the wildcard matcher.

Implemented (and added tests) for the DOS_DOT match symbol.  Finally, added the
ability to skip translation for the '?' match symbol so it can be tested
without getting translated.