Fixed VFS #Move bugs and added File.{Copy,Move}
Fixed an issue with the VFS #Move method which was not moving
file->file correctly.  The method also had zero tests.  New tests now
added to cover the move file->file case, but coverage is still limited.

Added FileObject code for Copy and Move operations.  Move is still in
progress but Copy is complete.