Fixed issue with BOM/encodings for append-only streams
When loading the contents of an existing ASCII file using the
FileObject.OpenAsTextStream(8, -1), the returned TextStream changed
the encoding of the existing characters from ASCII to UTF-16.  Had to
rework how file loading is done and add some hateful branching to test
whether or not we're in append-only mode to get it to "do the right
thing".  Very brittle code.  Seems to work for now, but will likely
turn up more bugs in the future.