Added 'finished' message and change event hooks
Event hooks are now called as soon as en event is emitted, rather than
after.  This means that reporters can now report events "live", which
really helps when analysing a new sample.

Events now end with a 'finished' message.

Restructured the reporters to use the new event structure.

Fixed a bug with the registry helper which was writing 'undefined'
instead of the folder path when loading from the CFG.